requirements.txt
Django>=5.1.7
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.1
Pillow>=10.2.0
python-dotenv>=1.0.0
redis>=5.0.1
gunicorn>=21.2.0
psycopg2-binary>=2.9.9  # For PostgreSQL support
whitenoise>=6.6.0  # For serving static files
django-storages>=1.14.2  # For cloud storage support 